backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
numpy
scipy
scikit-learn
lightgbm
joblib
pdfplumber
pytesseract
Pillow
google-generativeai
requests
python-dotenv
PyJWT
werkzeug
python-multipart
pyzbar
reportlab
